https://github.com/teamed/quiz.git
I added some java doc to the class and methods. 
Changed the i and o variable names for inputStream and outputStream
Included a parameter in the getContent method in order to indicate if
Unicode content should be included or not. The other method
(getContentWithoutUnicode was removed). 
I also added synchronized to the saveContent and getContent methods, and
closed the inputStream and outputStream in both methods.