Refactor Parser.java
These changes address a number of concurrency, memory, and general
efficiency issues.  Parser#file is now volatile rather than having the accessor
and mutator synchronized, as such synchronization is unnecessary and
could collide with not only with later methods added to this class, but also
with other classes using an instance of Parser's intrinsic lock.  In this case,
marking Parser#file as volatile avoids any potential pitfalls of intrinsic
locking as well as the cost(while admittedly miniscule) of obtaining the
intrinsic lock, while still providing the needed thread safety(atomic read
and write operations).

File IO continues to use ASCII(the format indicated by recognizing each
byte as an individual character), and is simplified through the use of Java
7 API's.  All file streams and handles are now appropriately closed.  File
reading that is not directly handled by Java 7 API's now uses buffering.
This buffering is especially significant when reading the file one byte
(or really, character) at a time.

The possibility of Parser reaching an illegal state was removed with a
validation check added to Parser#setFile(File).  By placing the validation
check in the accessor, errors can be caught when an invalid input is
provided, rather than when the standard operations all fall apart.