JimDeanSpivey quiz answer
1) renamed Parser.java to FileAccessor.java
-There was no parser, FileAccessor is a much more apt name.

2) changed how getContentWithoutUnicode works.
- There are many possible encodings that using data beyond 0x80 (eg:
  cyrillic). Added a boolean called "asciiOnly" instead since cahracters
bellow 0x80 are more commonly ASCII.

3) Re-did file reading to use the Reader convenience classes. There was
also a slight bug in the previous code that skipped reading that was a
null byte (0x00):
    while ((data = i.read()) > 0) { // <-- note it should be i.read() >=
0

4) Added a more performant saveContents method that uses an inputstrem
to point to larger amounts of data, in case all the data to be written
can't simply fit into a single string.

5) Decided to leave the test mainclass method. Since the point of this
is to evaluate coding, it is better to leave the test driving code
intact. Normally such code would not be submitted (or would be better
laid out in a proper unit test framework).