Quiz - Refactor Parse.java
Remove duplicate methods 'getContent()' and 'getContentWithoutUnicode()' and
use a single method 'getContent(boolean isUnicode)' instead.

Use of the StringBuilder instead string concatenation to achieve better performance.

Use of the Java 7 try-with-resources to achieve correct closing of the streams.

Rework creation of the streams to achieve proper reading/writing of the Unicode
characters like Cyrillic or Hebrew from/to the file.